+++ Core/MIPS/MIPSIntVFPU.cpp
@@ -826,7 +826,7 @@ namespace MIPSInt
     {
       switch ((op >> 16) & 0x1f)
       {
-      case 1: d[i] = currentMIPS->rng.R32(); break;  // vrndi - TODO: copy bits instead?
+      case 1: d[i] = (float)currentMIPS->rng.R32(); break;  // vrndi - TODO: copy bits instead?
       case 2: d[i] = 1.0f + ((float)currentMIPS->rng.R32() / 0xFFFFFFFF); break; // vrndf1   TODO: make more accurate
       case 3: d[i] = 2.0f + 2 * ((float)currentMIPS->rng.R32() / 0xFFFFFFFF); break; // vrndf2   TODO: make more accurate
       case 4: d[i] = 0.0f;  // Should not get here
